Fix: food objects properly combine onto plate
Finished combine script, created ObjectGrabCheck script (to check if an object is being grabbed or not)

Combinables can be combined onto a plate if user places directly onto plate, or if plate is dragged over object.

Edited cooking object prefabs (foods) to have ObjectGrabCheck script.

Modified scene so that quick tests can be done to see combinability in action.